+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ### Unreleased
 
 * Add initial spike for parsing, building and manipulating an in memory representation of the `system` data from Endless Sky. The method `zifnab.config.SystemConfig.from(DataElement)` is used to parse system data.
+* Add simplified `element()` and `comment()` factory methods onto both `DataDocument` and `DataElement` for creating children. Eaxample usage includes:
+  - `e.element(location, "planet", "mars")`
+  - `e.element("planet", "mars")`
+  - `e.comment(location, "Great planet!")`
+  - `e.comment("Great planet!")`
+
 
 ### [v0.02](https://github.com/realityforge/zifnab/tree/v0.02) (2019-08-09)
 [Full Changelog](https://github.com/realityforge/zifnab/compare/v0.01...v0.02)